Redefinition of structures
Changed structures to be more compact, putting colours and position
together in arrays. Also removed individual shape structures to be one
type, which will be defined by an enum declaring which shape. May add
unions for data in this shape struct. Also added a shape linked list.
Attempting to use stack memory for the main data struct as well as
ambience, camera and light but may have to change this to heap
allocation in future.